Date checking tricks
Added a check such that a goal is satisfiable iff there is at least one
measurement whose measuring date belong to the goal's valid range (i.e.
[creationDate, deadline]).

Added a check such that the user is advised to add a new measurement for
some measureType also if it exists a measurement for that measureType,
but its measuring date is before the corresponding active goal's
creation.